requirements.txt
Flask
Flask-Mail
Flask-Cors
firebase-admin
python-dotenv
gunicorn